Move luajit test to its own package.
(Assisted by Claude)
