+++ Makefile
@@ -1,5 +1,6 @@
 .PHONY: all all-common clean ebpf generate generate-collector test test-deps \
-	test-junit protobuf docker-image agent legal integration-test-binaries \
+	test-junit test-luajit-offsets protobuf docker-image agent legal \
+	integration-test-binaries \
 	codespell lint ebpf-profiler format format-ebpf format-go pprof-execs \
 	pprof_1_23 pprof_1_24 pprof_1_24_cgo otelcol-ebpf-profiler \
 	rust-components rust-targets rust-tests vanity-import-check vanity-import-fix \
@@ -125,6 +126,11 @@ vanity-import-fix:
 test: generate ebpf test-deps
 	# tools/coredump tests build ebpf C-code using CGO to test it against coredumps
 	CGO_ENABLED=1 go test $(GO_FLAGS) -tags $(GO_TAGS) ./...
+
+# Lives in its own module to keep heavy dependencies (testcontainers,
+# Docker, etc.) out of the main go.mod.
+test-luajit-offsets: generate
+	go test -C ./interpreter/luajit/offsetstest $(GO_FLAGS) -tags luajit_offsets_test ./...
 
 test-junit: generate ebpf test-deps
 	mkdir -p $(JUNIT_OUT_DIR)
